Add Linux-HA resource agent for Manila
The OpenStack HA guide [1][2] describes how to configure the Linux
clustering components Corosync and Pacemaker to run OpenStack services
as active/passive HA processes.

The Manila resource agent was modeled after the one for Cinder, plus
a few fixes needed to make it work with Devstack.  A separate commit
to the OpenStack HA Guide will add a section covering Manila.

[1] http://docs.openstack.org/high-availability-guide/content/ch_preface.html
[2] https://github.com/openstack/ha-guide

Change-Id: I14f9598cf682ec3c43c2163a3dfabf18a59d7baa