Remove Keystone dependency from neutron-server RA
When we check the availability of Neutron API service, now we simply
check the response code of "List API version" call instead of getting a
token from Keystone, then checking a Neutron API endpoint using that
token. This way we don't need a token anymore so the checking process
will not depend on the availability on Keystone.

Partial-Bug: #1540363
Change-Id: I5fee8d47bd8e9af9f415b9f74c4f9325ac99df2f
Signed-off-by: Norbert Illes <norbert.e.illes@ericsson.com>