Temporary ignore bashate E006 errors
The currently available bashate releases are considering heredocs as
normal code lines, hence lines longer than 79 columns in these sections
are also considered as E006 violations. As the OCF scripts are
containing lots of heredocs, we are affected by this behaviour.
However, there is a commit in the bashate repository (649c7dc79948)
which modifies bashate to ignore long lines in heredocs.

Currently there is no bashate release which contains the above commit,
so we ignore E006 errors until a new bashate released.

Change-Id: I33a9737ce1ec7eddab0b24ddedefe5c17da03b7a
Partial-Bug: #1550203
Signed-off-by: Norbert Illes <norbert.e.illes@ericsson.com>