De-CamelCase nova compute RAs
All the other RAs are lowercase and hyphenated, so make NovaCompute and
NovaEvacuate consistent with those.

Any distro-specific code which assumes the CamelCase versions can easily
have their packages tweaked accordingly to still install the files using
the old versions, or to install the old names as symlinks to the new
ones.  So this shouldn't cause any significant disruption.

Closes-Bug: #1557666
Change-Id: I9c18cbe45ea2559611e4571b8ee378ea0ae3723f