[PoC] Call masakari APIs from a resource agent
This is a PoC of a resource agent that calls masakari APIs if
instance's host failure is occurred.
The purposes of this patch is below:
- Show how to call the masakari APIs from nova-host-alerter.
- Indicate the implementation of the masakari driver.

This is just a PoC based on an idea called as "Modular architecture".

On this patch, we assumed that nova-host-alerter has 'driver' param
in the primitive definition, and that 'masakari' or 'mistral' is
set for the driver param.
And we plan to place masakari_driver.py and masakari_driver.conf
in the same directory as nova-host-alerter.

reference:
* The idea of modular architecture
  https://aspiers.github.io/openstack-day-israel-2017-compute-ha/#/nova-host-alerter
  https://aspiers.github.io/openstack-day-israel-2017-compute-ha/#/modular
* Specs of a method to recover all virtual machines
  https://github.com/openstack/openstack-resource-agents-specs/blob/master/specs/newton/approved/newton-instance-ha-host-monitoring-spec.rst
  https://github.com/openstack/openstack-resource-agents-specs/blob/master/specs/newton/approved/newton-instance-ha-host-recovery.rst
  https://review.openstack.org/#/c/406659/

Change-Id: I6768a1822ed5f19bc66f0d6d6887194bbc32abad
Co-Authored-By: Kengo Takahara <takahara-kn@njk.co.jp>
Story: 2002124
Task: 19803